requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0
tabulate>=0.9.0
schedule>=1.2.0
